requirements.txt
Django==5.2.13
django-environ==0.13.0
djangorestframework==3.17.1
